requirements.txt
openpyxl
PyAutoGUI
pyperclip
selenium
pynput
pytest
pillow
pyautogui
opencv-python
allure-pytest
pytest-dependency
pytest-rerunfailures